Hold the rules a walk reads a program by apart from the check that reads them
Entering a binding, taking a condition as holding, and reading what a type
guarantees are answers about a program. The question they add up to — can
anything at all stand here — has more than one reader, and the only one that
could answer it was a private step of the invariant-discharge walk, which
consumes it by returning.

So they are held in PathEngine and the check keeps one. Nothing about what is
read changes: the same objects are built once and named where the check reads
them, and the seeding, the binding and the assumption run the code they ran.
Written now so that the reader added next asks these rules rather than deriving
a second set, which would part company at the first clause a domain other than
the numbers has a word for.

Behaviour preserving. souther-compiler 5077 tests green, and the ten
souther-examples modules answer `examples --format json` byte for byte as they
did before.
